+++ src/lib/components/PageContainer/index.styled.js
@@ -1,19 +1,8 @@
 import styled from 'styled-components'
 import { Typography } from 'antd'
 
-import Card from '../Card'
 import Tabs from '../Tabs'
 import { COLORS } from '../../constants'
-
-export const Header = styled(Card)`
-  border-radius: 0 0 6px 6px;
-  margin-bottom: 8px;
-  background: ${COLORS.white} !important;
-
-  .ant-card-body {
-    padding: 12px 40px;
-  }
-`
 
 export const Title = styled(Typography.Title)`
   margin-top: 6px;